requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
packaging==23.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.5
whitenoise==6.5.0
